Tweaks and fixes to enemyPatroller behaviour and animations. Attack functionality tested and fixed on enemyPatroller. Player relocation to prison marker added to successful enemy melee attacks. Modified lootSpawnArea script, to parent all the loot on the spawn area transform. Tweaks to enemyPatroller material (GuardTex) to reduce shininess.

+++ CastleRaid/Assets/Scripts/Enemy/EnemyAnimationController.cs
@@ -99,7 +99,7 @@ public class EnemyAnimationController : MonoBehaviour
     {
         if (isActive)
         {
-            animator.SetInteger("MovementSpeed", newSpeed);
+            animator.SetFloat("MovementSpeed", newSpeed);
         }
     }
     #endregion
